+++ clients/admin-ui/src/features/common/nav/nav-config.test.ts
@@ -255,6 +255,10 @@ describe("configureNavGroups", () => {
           { title: "Organization", path: routes.ORGANIZATION_MANAGEMENT_ROUTE },
           { title: "Taxonomy", path: routes.TAXONOMY_ROUTE },
           { title: "Email templates", path: routes.EMAIL_TEMPLATES_ROUTE },
+          {
+            title: "Messaging providers",
+            path: routes.MESSAGING_PROVIDERS_ROUTE,
+          },
           { title: "About Fides", path: routes.ABOUT_ROUTE },
         ],
       });
@@ -287,6 +291,10 @@ describe("configureNavGroups", () => {
           {
             title: "Email templates",
             path: routes.EMAIL_TEMPLATES_ROUTE,
+          },
+          {
+            title: "Messaging providers",
+            path: routes.MESSAGING_PROVIDERS_ROUTE,
           },
           { title: "About Fides", path: routes.ABOUT_ROUTE },
         ],